feat: replace use of illegal_variable_name() with invalid_variable_name() and consolidate
