+++ docassemble_base/docassemble/base/parse.py
@@ -17,7 +17,6 @@ import tempfile
 import json
 import platform
 import textwrap
-from keyword import iskeyword
 from urllib.request import urlretrieve
 from io import StringIO
 from collections import abc, OrderedDict, namedtuple
@@ -53,7 +52,7 @@ import docassemble.base.file_docx
 from docassemble.base.error import DAError, DANotFoundError, MandatoryQuestion, DAErrorNoEndpoint, DAErrorMissingVariable, ForcedNameError, QuestionError, ResponseError, BackgroundResponseError, BackgroundResponseActionError, CommandError, CodeExecute, DAValidationError, ForcedReRun, LazyNameError, DAAttributeError, DAIndexError, DAException, DANameError, DASourceError
 import docassemble.base.functions
 import docassemble.base.util
-from docassemble.base.functions import pickleable_objects, word, get_language, RawValue, get_config
+from docassemble.base.functions import pickleable_objects, word, get_language, RawValue, get_config, invalid_variable_name
 from docassemble.base.logger import logmessage
 from docassemble.base.pandoc import MyPandoc
 from docassemble.base.mako.template import Template as MakoTemplate
@@ -96,7 +95,6 @@ match_inside_brackets = re.compile(r'\[(.+?)\]')
 match_brackets = re.compile(r'(\[.+?\])')
 match_brackets_or_dot = re.compile(r'(\[.+?\]|\.[a-zA-Z_][a-zA-Z0-9_]*)')
 complications = re.compile(r'[\.\[]')
-invalid_variable_name_chars = re.compile(r'[\x00-\x1f\x7f-\x9f\u200b-\u200d\u202a-\u202e\u2060\ufeff+\-*/&|^<>=!{}();,@#\\]')
 list_of_indices = ['i', 'j', 'k', 'l', 'm', 'n']
 extension_of_doc_format = {'pdf': 'pdf', 'docx': 'docx', 'rtf': 'rtf', 'rtf to docx': 'docx', 'tex': 'tex', 'html': 'html', 'md': 'md', 'raw': 'raw'}
 DO_NOT_TRANSLATE = """<%doc>
@@ -7991,20 +7989,6 @@ def restore_backup_vars(the_user_dict, backups):
         the_user_dict[var] = val
 
 
-def illegal_variable_name(varname: str) -> bool:
-    if invalid_variable_name_chars.search(varname):
-        return True
-    try:
-        tree = ast.parse(varname)
-    except SyntaxError:
-        return True
-    allowed_nodes = (ast.Module, ast.Expr, ast.Name, ast.Attribute, ast.Constant, ast.Subscript, ast.Slice, ast.Load)
-    for node in ast.walk(tree):
-        if not isinstance(node, allowed_nodes):
-            return True
-    return False
-
-
 def double_to_single(text):
     if text.startswith('[') and text.endswith(']'):
         try:
@@ -8825,13 +8809,13 @@ class Interview:
                     if interview_status.current_info['action'] in ('_da_list_remove', '_da_list_add', '_da_list_complete'):
                         for the_key in ('list', 'item', 'items'):
                             if the_key in interview_status.current_info['arguments']:
-                                if illegal_variable_name(interview_status.current_info['arguments'][the_key]):
+                                if invalid_variable_name(interview_status.current_info['arguments'][the_key]):
                                     raise DASourceError("Invalid name " + interview_status.current_info['arguments'][the_key])
                                 interview_status.current_info['action_' + the_key] = eval(interview_status.current_info['arguments'][the_key], user_dict)
                     if interview_status.current_info['action'] in ('_da_dict_remove', '_da_dict_add', '_da_dict_complete'):
                         for the_key in ('dict', 'item', 'items'):
                             if the_key in interview_status.current_info['arguments']:
-                                if illegal_variable_name(interview_status.current_info['arguments'][the_key]):
+                                if invalid_variable_name(interview_status.current_info['arguments'][the_key]):
                                     raise DASourceError("Invalid name " + interview_status.current_info['arguments'][the_key])
                                 interview_status.current_info['action_' + the_key] = eval(interview_status.current_info['arguments'][the_key], user_dict)
                 # else:
@@ -10367,21 +10351,6 @@ def ensure_object_exists(saveas, datatype, the_user_dict, commands=None):
         for command in commands:
             # logmessage("Doing " + command)
             exec(command, the_user_dict)
-
-
-def invalid_variable_name(varname: str) -> bool:
-    if not isinstance(varname, str):
-        return True
-    if invalid_variable_name_chars.search(varname):
-        return True
-    varname_segments = varname.split(".")
-    for segment in varname_segments:
-        first, sep, _ = segment.partition("[")
-        if not first.isidentifier() or iskeyword(first):
-            return True
-        if sep:
-            return illegal_variable_name(segment)
-    return False
 
 
 def exec_with_trap(the_question, the_dict, old_variable=None):
